requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib
python-dotenv
openai
plotly

